requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
Flask==2.3.3
Flask-Cors==4.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
lxml==5.4.0
MarkupSafe==3.0.2
nltk==3.8.1
numpy==2.2.6
openai==1.82.0
packaging==24.2
plotly==5.24.1
portalocker==3.1.1
proto-plus==1.26.1
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dotenv==1.0.0
pywin32==310
regex==2024.11.6
requests==2.32.3
rouge_score==0.1.2
rsa==4.9.1
sacrebleu==2.3.1
six==1.17.0
sniffio==1.3.1
tabulate==0.9.0
tenacity==9.0.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
